refactor: integrate pronoun distribution as collapsible in AttendeeList
- Move pronoun distribution from standalone component into AttendeeList
- Add toggle button "Pronoun Distribution" with expand/collapse
- Data is lazily fetched only when user clicks to expand
- Remove standalone PronounDistribution.svelte component
- Cleaner UX: related info stays together, one less card in sidebar
